add binary lib
