+++ cloud-libs/util/binary/eslint.config.cjs
@@ -1,0 +1,3 @@
+const baseConfig = require('../../../eslint.config.js');
+
+module.exports = [...baseConfig];
