+++ cloud-libs/util/binary/project.json
@@ -1,0 +1,9 @@
+{
+  "name": "binary",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "cloud-libs/util/binary/src",
+  "projectType": "library",
+  "tags": [],
+  "// targets": "to see all targets run: nx show project binary --web",
+  "targets": {}
+}
