+++ libs/features/service-protocol/src/lib/entries/attachInvocation.ts
@@ -1,4 +1,4 @@
-import {
+import type {
   Invocation,
   JournalEntryV2,
 } from '@restate/data-access/admin-api/spec';
