+++ tsconfig.base.json
@@ -87,6 +87,7 @@
       "@restate/ui/split-button": ["libs/ui/split-button/src/index.ts"],
       "@restate/ui/table": ["libs/ui/table/src/index.ts"],
       "@restate/ui/tooltip": ["libs/ui/tooltip/src/index.ts"],
+      "@restate/util/binary": ["cloud-libs/util/binary/src/index.ts"],
       "@restate/util/errors": ["libs/util/errors/src/index.ts"],
       "@restate/util/feature-flag": ["libs/util/feature-flag/src/index.ts"],
       "@restate/util/humantime": ["libs/util/humantime/src/index.ts"],
